stats-sleeper/requirements.txt
sleeper-api-wrapper>=1.0.4
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
tabulate>=0.9.0
jinja2>=3.1.0
